Fixing
Fixed a lotta stuff
