+++ src/object_classes/food_object_eaten.js
@@ -6,10 +6,16 @@ import AsyncStorage from "@react-native-async-storage/async-storage";
 // This class also contains functions to save the object locally.
 // Additional functions can be added as needed.
 
-const prefix = '@Food:';
+const prefix = '@Eaten_Food:';
 
-export class food_object {
-        constructor(name, calories, carbs, sugar, fat, protein, salt) {
+export class food_object_eaten {
+        constructor(date, amount, name, calories, carbs, fat, protein) {
+        if (typeof date !== 'string') {
+            throw new Error('Date must be a string');
+        }
+        if (typeof amount !== 'number') {
+            throw new Error('Amount must be a number');
+        }
         if (typeof name !== 'string') {
             throw new Error('Name must be a string');
         }
@@ -19,32 +25,26 @@ export class food_object {
         if (typeof carbs !== 'number') {
             throw new Error('Carbs must be a number');
         }
-        if (typeof sugar !== 'number') {
-            throw new Error('Sugar must be a number');
-        }
         if (typeof fat !== 'number') {
             throw new Error('Fat must be a number');
         }
         if (typeof protein !== 'number') {
             throw new Error('Protein must be a number');
         }
-        if (typeof salt !== 'number') {
-            throw new Error('Salt must be a number');
-        }
 
+        this.date = date;
+        this.amount = amount;
         this.name = name;
         this.calories = calories;
         this.carbs = carbs;
-        this.sugar = sugar;
         this.fat = fat;
         this.protein = protein;
-        this.salt = salt;
     }
 
     async saveLocal() {
         try {
             const jsonValue = JSON.stringify(this);
-            var id = prefix + this.name;
+            var id = prefix + this.name + ":" + this.date; // @EatenFood:date:name
             await AsyncStorage.setItem(id, jsonValue);
             AsyncStorage.getItem(this.name).then((res) => console.log(res))
         }
